Change the sngl_inspiral event_id element from UINT4 to long. Original: ce45e897c5f0f390e04816e2cbcf1cfa3703c287

+++ lal/packages/tools/src/SnglBurstUtils.c
@@ -126,7 +126,7 @@ void XLALFreeSnglBurst(SnglBurstTable *event)
 void XLALSnglBurstAssignIDs(SnglBurstTable *head)
 /* </lalVerbatim> */
 {
-	unsigned int id;
+	long id;
 
 	for(id = 0; head; head = head->next)
 		head->event_id = id++;
